Rename file containing SBP callbacks
