+++ package/sbp_settings_daemon/sbp_settings_daemon/src/main.c
@@ -12,7 +12,8 @@
 
 #include <libpiksi/sbp_pubsub.h>
 #include <libpiksi/logging.h>
-#include "settings.h"
+
+#include "settings_sbp_cb.h"
 
 #define PROGRAM_NAME "sbp_settings_daemon"
